Fix Issue With Extra Strings Erroring
If a String Literal is the last token to be found in the file, but
is a child of a STRINGZ, then previously this would cause errors as
a String is not a valid starting token.

This was due to the way Iter::position works, and so this now repeatedly
consumes from the tokens until it finds a token that isn't a string --
if it happens to reach the end of the queue, it will just stop
processing.